+++ src/MeasureMap/ProfilerSettingsExtensions.cs
@@ -22,5 +22,15 @@ namespace MeasureMap
                 _ => WorkerThread.Factory
             };
         }
+
+        /// <summary>
+        /// Create a new ExecutionContext based on the Settings. This can be used in the OnStartPipeline Event to create the IExecutionContext for the Pipeline
+        /// </summary>
+        /// <param name="settings"></param>
+        /// <returns></returns>
+        public static IExecutionContext CreateContext(this ProfilerSettings settings)
+        {
+            return new ExecutionContext(settings);
+        }
     }
 }